test/acceptance/features/requirements.txt
behave==1.2.6
requests==2.27.1
polling2==0.5.0
PyYAML==6.0
semver==2.13.0
cryptography==37.0.2
kubernetes==25.3.0
